Version 1.2 complete
Translation function added. (top right button)

To make it work, I decided to use external language files, which means that the program needs to be installed correctly as indicated in the download page.
To make it easier for everyone, I plan to translate this app in C++ in the future, to create a .exe installer.

The very last thing I now want this app to be able to do, is to work as well with negative numbers.